INFESTED INNOVATION RELEASE 1.0.0
INFESTED INNOVATION OFFICIAL RELEASE 1.0.0.

ABSTRACT:

A modpack that overhauls the core of Minecraft and repurposes it into a survival frenzy. You must build from nothing and learn to sustain yourself to survive against the ever-strengthening hordes of mobs. This modpack requires innovative structural designs as the new mobs have abilities never seen before. Furthermore, you must innovate using newly founded resources from Tinker's Construct and Create to advance yourself and your allies so that you may take on new adventures and place foot where no ordinary man dares to. Your mission in this modpack, is to not only survive, but to learn how to thrive against the numerous dangers within this modpack.

FEATURES:

Below is a list of features implemented into this modpack and meticulously configured and changed to suit the purpose of this new gamemode:

– All New Structure Design and Spawning.
– Custom Day/Night Cycle for Easier Adventuring.
– Customized Recipes for Better Scaling With Modpack Difficulty.
– Harder Mobs That Continue To Advance.
– Loads of Utility, QoL, & Performance Mods for Smoother Gameplay.
– Mind-Boggling Contraptions Using Create.
– New Tools and Armor Using TInker’s Construct.
– Overhauled Enchanting System Using New Enchanting Infuser Block.
– Realistic Temperature for More Challenging Expeditions to the Nether.
– Skillbook To Upgrade Character and Provide Unique Gameplay.
– Various Assortments of Foods Using Pam’s Harvest Craft.
– And So So Much More!

CHANGELOG:

This official release has gone through several changes since the previous committed pack. Here's a list of the changes made:

– Added Various Performance Mods.
– Configured Day/Night Cycle for Less Aggressive Difficulty Scaling.
– Made Structures Less Sparse.
– Minor Reworks to Mobs and Loot Tables.
– New Create Mods for Better Automation.
– Removed Dynamic Trees To Make Wood Farms Possible.
– Removed Unnecessary Mods.
– Updated MANY Mods to Latest Stable Version.

Have Fun!